Added Python script to build Rupaya binaries for different platforms using Gitian descriptors
This commit adds a new Python script to build Rupaya binaries for different platforms using Gitian descriptors. The script uses the argparse module to accept command-line arguments for the GitHub organization and project, as well as the platforms to build. It also includes functions to get the list of tags from the Rupaya GitHub repository, prompt the user to select a tag, check if required files are present for OSX build, and build a platform for a specific tag.

The default values for the GitHub organization and project, platforms to build, Docker image name, cache directory, and result directory are now defined as constants. The script also uses the requests module to get the list of tags from the Rupaya GitHub repository, and the subprocess module to run Docker commands.

Overall, this commit adds a useful Python script for building Rupaya binaries for different platforms using Gitian descriptors, which can be used to improve the build process for the Rupaya project.